(WIP) Improve tree component
