+++ src/Tree/ChildTree.tsx
@@ -1,9 +1,6 @@
 import * as React from "react"
-
-import Icon from "../Icon/Icon"
-import NameTag from "../NameTag/NameTag"
-import { Container, DeleteNode, Header, Label, TreeIcon } from "./styledComponents"
 import Tree, { TreeProps } from "./Tree"
+import TreeItem, { Container } from "./TreeItem"
 
 type Props = TreeProps["trees"][-1]
 
@@ -16,45 +13,38 @@ const ChildTree: React.SFC<Props> = ({
   disabled,
   forwardRef,
   childNodes = [],
-  onClick: onNodeClick,
+  onClick,
   onRemove,
   ...props
 }) => {
   const [isOpen, setIsOpen] = React.useState(Boolean(initiallyOpen))
-  const toggle = (e: React.MouseEvent<HTMLDivElement>) => {
-    e.stopPropagation()
-    if (childNodes && childNodes.length) {
-      setIsOpen(!isOpen)
-    }
-  }
-
   const hasChildren = Boolean(childNodes && childNodes.length)
+  const onNodeClick =
+    !disabled && (hasChildren || onClick)
+      ? (e: React.MouseEvent<HTMLDivElement>) => {
+          e.stopPropagation()
+          if (hasChildren) {
+            setIsOpen(!isOpen)
+          }
+          if (onClick) {
+            onClick()
+          }
+        }
+      : undefined
 
   return (
     <Container ref={forwardRef} disabled={Boolean(disabled)} hasChildren={hasChildren} {...props}>
-      <Header
-        onClick={e => {
-          toggle(e)
-          if (onNodeClick) {
-            onNodeClick()
-          }
-        }}
+      <TreeItem
+        onNodeClick={onNodeClick}
         highlight={Boolean(highlight)}
-      >
-        {hasChildren && <TreeIcon color="color.text.lightest" size={12} left name={isOpen ? "ChevronDown" : "Add"} />}
-        {!hasChildren && tag && (
-          <NameTag condensed left color={color}>
-            {tag}
-          </NameTag>
-        )}
-        <Label hasChildren={hasChildren}>{label}</Label>
-        {onRemove && (
-          <DeleteNode onClick={onRemove}>
-            <Icon size={12} name="No" />
-          </DeleteNode>
-        )}
-      </Header>
-      {childNodes && isOpen && <Tree trees={childNodes} />}
+        hasChildren={hasChildren}
+        isOpen={isOpen}
+        tag={tag}
+        label={label}
+        color={color}
+        onRemove={onRemove}
+      />
+      {hasChildren && isOpen && <Tree trees={childNodes} />}
     </Container>
   )
 }
